+++ server/src/eu/thomaskuenneth/souffleur/Server.java
@@ -1,20 +1,24 @@
 package eu.thomaskuenneth.souffleur;
 
-import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpHandler;
-import com.sun.net.httpserver.HttpServer;
+import com.sun.net.httpserver.*;
 
-import javax.imageio.ImageIO;
+import javax.net.ssl.*;
 import java.awt.AWTException;
 import java.awt.Robot;
-import java.awt.image.BufferedImage;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.net.URI;
+import java.security.*;
+import java.security.cert.CertificateException;
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
+import java.time.format.FormatStyle;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+
 
 public class Server implements HttpHandler {
 
@@ -22,14 +26,13 @@ public class Server implements HttpHandler {
     public static final String NEXT = "next";
     public static final String PREVIOUS = "previous";
     public static final String HOME = "home";
-    public static final String QRCODE = "qrcode";
     public static final String HELLO = "hello";
     private static final Logger LOGGER = Logger.getLogger(Server.class.getName());
 
     private final Robot robot;
     private final ServerCallback callback;
 
-    private HttpServer httpServer;
+    private HttpsServer httpServer;
     private String address;
     private int port;
     private String secret;
@@ -65,27 +68,56 @@ public class Server implements HttpHandler {
                 callback.commandReceived(END);
                 sendStatus(t, 200);
             }
-            case QRCODE -> {
-                sendQRCode(t);
-            }
             case HELLO -> {
-                sendStringResult(t, "Hello, world!");
+                sendStringResult(t, String.format("Hello, world! It's %s.",
+                        DateTimeFormatter.ofLocalizedTime(FormatStyle.SHORT).format(LocalTime.now())));
                 callback.commandReceived(HELLO);
             }
+            default -> sendStatus(t, 404);
         }
     }
 
-    public void start(String address, int port, String secret) throws IOException {
-        InetAddress inetAddress = InetAddress.getByName(address);
-        InetSocketAddress socketAddress = new InetSocketAddress(inetAddress, port);
-        this.address = address;
-        this.port = port;
-        this.secret = secret;
-        this.httpServer = HttpServer.create(socketAddress, 0);
-        this.httpServer.createContext("/souffleur/" + secret, this);
-        this.httpServer.setExecutor(null);
-        this.httpServer.start();
-        LOGGER.log(Level.INFO, getQRCodeAsString());
+    public boolean start(String address, int port, String secret) {
+        boolean success = false;
+        try {
+            InetAddress inetAddress = InetAddress.getByName(address);
+            InetSocketAddress socketAddress = new InetSocketAddress(inetAddress, port);
+            this.address = address;
+            this.port = port;
+            this.secret = secret;
+            this.httpServer = HttpsServer.create(socketAddress, 0);
+            SSLContext sslContext = SSLContext.getInstance("TLS");
+            char[] password = "password".toCharArray();
+            KeyStore ks = KeyStore.getInstance("JKS");
+            InputStream is = getClass().getResourceAsStream("/souffleur.jks");
+            ks.load(is, password);
+            KeyManagerFactory kmf = KeyManagerFactory.getInstance("SunX509");
+            kmf.init(ks, password);
+            TrustManagerFactory tmf = TrustManagerFactory.getInstance("SunX509");
+            tmf.init(ks);
+            sslContext.init(kmf.getKeyManagers(), tmf.getTrustManagers(), null);
+            this.httpServer.setHttpsConfigurator(new HttpsConfigurator(sslContext) {
+                @Override
+                public void configure(HttpsParameters params) {
+                    SSLContext c = getSSLContext();
+                    SSLEngine engine = c.createSSLEngine();
+                    params.setNeedClientAuth(false);
+                    params.setCipherSuites(engine.getEnabledCipherSuites());
+                    params.setProtocols(engine.getEnabledProtocols());
+                    SSLParameters sslParameters = c.getDefaultSSLParameters();
+                    params.setSSLParameters(sslParameters);
+                }
+            });
+            this.httpServer.createContext("/souffleur/" + secret, this);
+            this.httpServer.setExecutor(null);
+            this.httpServer.start();
+            LOGGER.log(Level.INFO, getQRCodeAsString());
+            success = true;
+        } catch (IOException | NoSuchAlgorithmException | KeyManagementException | KeyStoreException |
+                 UnrecoverableKeyException | CertificateException e) {
+            LOGGER.log(Level.SEVERE, null, e);
+        }
+        return success;
     }
 
     public void stop() {
@@ -96,18 +128,7 @@ public class Server implements HttpHandler {
     }
 
     public String getQRCodeAsString() {
-        return String.format("http://%s:%s/souffleur/%s/", address, port, secret);
-    }
-
-    private void sendQRCode(HttpExchange t) {
-        BufferedImage image = Utils.generateQRCode(getQRCodeAsString());
-        try (OutputStream os = t.getResponseBody()) {
-            t.sendResponseHeaders(200, 0);
-            ImageIO.write(image, "jpg", os);
-            t.getResponseHeaders().add("Content-Type", "image/jpeg");
-        } catch (IOException e) {
-            LOGGER.log(Level.SEVERE, "sendQRCode()", e);
-        }
+        return String.format("https://%s:%s/souffleur/%s/", address, port, secret);
     }
 
     private void sendStringResult(HttpExchange t, String text) {
@@ -122,7 +143,7 @@ public class Server implements HttpHandler {
     }
 
     private void sendStatus(HttpExchange t, int status) {
-        try (OutputStream os = t.getResponseBody()) {
+        try (OutputStream ignored = t.getResponseBody()) {
             t.sendResponseHeaders(status, 0);
         } catch (IOException e) {
             LOGGER.log(Level.SEVERE, "sendStringResult()", e);