+++ server/whatsnew.md
@@ -1,8 +1,17 @@
 # What's new?
 
+## 1.0.5
+- Using https
+- Returning 404 upon unrecognized requests
+- Removed *qrcode* message
+
 ## 1.0.4
 
 - first version with this file :-)
 - We are using Material-UI-Swing by Vincenzo Palazzo
 - Smiley indicator shows we received a hello
 - hello closes the QR code dialog
+
+## Earlier versions
+
+...this file wasn't there... ;-)